Add new property in DrawerController (#85)
DrawerController thinks it should present as UIPopoverPresentationController when the horizontal size is compact. But for larger phones like iPhone 11 Pro Max even in portrait orientation, horizontal size is regular.
(1) make sure in iPhone the vertical drawers are presented in slider mode.
(2) add a new property that clients can force the slider to be drawn in preferred content width in landscape in compact horizontal size.