Don't dismiss vertical drawers when the origin is fixed (#96)
We currently do NOT dismiss horizontal drawers upon device rotation, because they have a fixed origin (sides) so they will be presented from the same origin in any orientation or screen size. However, we DO dismiss vertical drawers with fixed origin, even though in most scenarios they will also be presented with the same origin for any device orientation. I believe we do this because there is one exception, when switching horizontal size class on iPad, we change the drawer for a popover and vice versa.

Let's handle that one exception, and dismiss the drawer, but allow vertical drawers NOT to be dismissed if we don't have to. This will help mobile clients that do not want the drawer to be dismissed upon device rotation, for instance, when the user had data input that can be lost.

We cannot avoid dismissal when using custom origins, because those origins can't be translated into a new frame, unless a complex coordinates logic is applied with multiple edge cases. If clients use custom origins, it'll be easier for them to persist the drawer, listen for orientation changes, and present it again, if that is absolutely required.

iPad is an interesting scenario. Popovers actually do not dismiss upon rotation or size change (even before this change). I think this is out of luck because of how the native popover controller behaves, but it will be buggy when the origin coordinates are in edge cases.