Change surfaceTertiary dark elevated color back to gray 900.
We need it to contrast with its non-elevated version to distinguish between a presented window and the window behind on iPhone. It sounds counterintuitive to select the same color, but iOS applies a dim to the window behind, which makes the current gray800 color look like gray900, which doesn't make a contrast.