Card component (#222)
* Card Component
Card has an icon, a title and an optional subtitle. The content can be arranged in a vertical or horizontal layout.

Due to a design misunderstanding, we are now separating the cards into this component, a PreviewCard component that will look like this component but with a preview image on top, and an announcement component. This Card will be used as a Quick Action card as part of the iPad redesign work

* add support for size category changes
* dynamic text and addressed comments
* Adjust labels setup to MSFLabel