Fix documentation in AvatarView + regression test
When importing FluentUI as a module into an ObjC file (via the @import command), if CLANG_WARN_DOCUMENTATION_COMMENTS is set to YES, it complains that the documentation is wrong.

This is a simple documentation fix to the source, but also adds a regression test. Sadly adding the first ObjC code to the project only for this specific regression test so that it doesn't break in the future.

Adding three settings to the unit test target:
CLANG_ENABLE_MODULES - make it easy to import FluentUI as a module in ObjC.
CLANG_WARN_DOCUMENTATION_COMMENTS - turns on the warning that we offended
GCC_TREAT_WARNINGS_AS_ERRORS - make sure if we break the warning, it gets red flagged