update segment control button accessibility hint (#307)
* update segment control button accessibility hint
We shouldn't be adding accessibility hints like "tap to go to something" in our library.
Interacting with a button with Voiceover is technically double tap. However, we really shouldn't be customizing string to tell users how to interact with our components because depending on the assistive technology that customer is using, how he or she activates a button might be different.
Instead, lets be consistent with our pillbuttonbar control which says the index of the button out of count of all buttons (similar to our tab bar works).