Partial fix for disappearing link
In the test app, when pressing "Link with overridden Target/Action", the link disappears.  It reappears when hovered over.  However, if `showsUnderlineWhileMouseInside` is `false`, the link doesn't reappear.

This seems related to the modal dialog: it doesn't occur if the custom action simply `print`s to the console.  It also doesn't occur if `isBordered` is `true`.  I couldn't figure out how to prevent the title from disappearing at all, but with this change it will automatically reappear when the modal dialog is dismissed.

Note that `updateTitle()` alone doesn't necessarily work: in some cases setting `attributedTitle` has no effect.  Resetting the underline style seems to be enough to ensure the title is redrawn properly.