Button test pane updates
Reduced view to 3 columns for Title only, Title & Image, and Image only.  Added a pop-up menu to set `imagePosition` for all buttons.

This allows us to more clearly see the effect of each `imagePosition` without having to consult the test code, and exercises the cases where we override `imagePosition` in the `titleRect` and `imageRect` methods for buttons that are missing one or the other.