Fix for 'pressed' state test
Following the Button Action fix the 'pressed' option in the 'Button State' pop-up was no longer effective in the test app.  Modified this test to directly set the state var rather than sending an event.  Also renamed the state var from 'mouseDown' to 'isPressed' so that it is distinct from the 'mouseDown' event callback.