Button Test Pane Improvements (#384)
* Minor refactoring
Adopted Swift idioms for NSRect and NSSize.  Simplified `drawingRect` method slightly, in the same vein as recent changes to `titleRect` and `imageRect`.

* Button test pane updates
Reduced view to 3 columns for Title only, Title & Image, and Image only.  Added a pop-up menu to set `imagePosition` for all buttons.

This allows us to more clearly see the effect of each `imagePosition` without having to consult the test code, and exercises the cases where we override `imagePosition` in the `titleRect` and `imageRect` methods for buttons that are missing one or the other.

* Bug fix
Detect and remove a transparent pixel that throws off alignment.

* Fix for 'pressed' state test
Following the Button Action fix the 'pressed' option in the 'Button State' pop-up was no longer effective in the test app. Modified this test to directly set the state var rather than sending an event. Also renamed the state var from 'mouseDown' to 'isPressed' so that it is distinct from the 'mouseDown' event callback.