Expose autocorrectionType as part of search bar
There are certain scenarios, where we don't want to use the built in autocorrection type. For example, when the search service provides us back suggestions, we'd rather use those since they might be more relevant. Additionally, searching for people names often misses auto-correct, so we'd like the ability to turn it off entirely.