Fix sizing issues with SearchBar
Without having an intrinsticContentSize, if you don't manually put a size on the search bar, then it ends up being a height of 0. In this case, you can't tap on the search bar even though it displays.

Validated in the test app that the new search bar test code works appropriately as well as in the navigation view in the multiple sets. I saw no changes to the sizes or pixel counts.