Toadgee/searchbarautocorrect (#427)
* Expose autocorrectionType as part of search bar

There are certain scenarios, where we don't want to use the built in autocorrection type. For example, when the search service provides us back suggestions, we'd rather use those since they might be more relevant. Additionally, searching for people names often misses auto-correct, so we'd like the ability to turn it off entirely.

* Adds simple SearchBarDemoController to test SearchBar

This adds a search bar demo controller to explicitly test the search bar autocorrect scenarios. More tests can be added.

Most tests for the search bar are in the NavigationControllerDemoController, but there's value in just testing the individual pieces.

* Order Demos folder alphabetically

* Missed one alphabetizing sort change in the project

* Fix sizing issues with SearchBar

Without having an intrinsticContentSize, if you don't manually put a size on the search bar, then it ends up being a height of 0. In this case, you can't tap on the search bar even though it displays.

Validated in the test app that the new search bar test code works appropriately as well as in the navigation view in the multiple sets. I saw no changes to the sizes or pixel counts.

* Remove unnecessary spacers from the SearchBarDemoController

Now that the SearchBar knows what size it should be, the layout happens wonderfully without these hacky extra spacers (this should have been my first hint that something was wrong)

Co-authored-by: Todd Harris <toddha@microsoft.com>