Update Segmented Control if statements and layoutBackgroundView
Changed some exhaustive if statements into switch statements. Removed layoutBackgroundView in favor of just setting the corner radius of the pill styles at init. Since the height of the pill buttons should always be 32 it should be safe to just set the corner radius to 16 and leave it at that, rather than setting it to 16 every time layoutSubviews is called.