Update Segmented Control selection animation (#466)
* Update Segmented Control selection animation

Changed the animation of the segmented control to use a mask layer to change the color of the text as the selection view passes a label. As part of that, removed the usage and dependency on Pill Button.

* Update where Segmented Control activates constraints

Moved the activation of constraints to a function that gets called one time during init. Also added a comment to explain the hierarchy of the pill styles

* Update Segmented Control if statements and layoutBackgroundView

Changed some exhaustive if statements into switch statements. Removed layoutBackgroundView in favor of just setting the corner radius of the pill styles at init. Since the height of the pill buttons should always be 32 it should be safe to just set the corner radius to 16 and leave it at that, rather than setting it to 16 every time layoutSubviews is called.

* Update Segmented Control constant name

Rename the pillHorizontalInset constant to pillContainerHorizontalInset to be more clear about how it's being used.

* Actually update Segmented Control constant name

I updated the constant name, but not where it was used. Oops.