add accessibilityHint for tabbaritemview for iOS 14 (#488)
We would like VoiceOver to read out "Home, 1 of 3, selected" when the VO focus is on a tabbar. However, right now it just says "Home, selected" in iOS 14 device.

This change is a partial revert of commit 1e17860 : it appears to be that .tabbar accessibilityTraits no longer automatically append index of the subview in iOS 14.
This change further improves sideTabBar experience so that voiceover focus order would be vertical navigation within the sidetabbar before it goes to other views. Also, VoiceOver will announce the index of the tabbaritemview.