Add Floating Action styles to vNext button
Added two new styles to the button, accent floating and subtle floating.

We don't have specs for a medium floating button, but the normal button has a medium size, so I made the medium floating buttons mostly use the small button specs and hid them in the demo app. Similarly, the guidance is to not show disabled floating buttons so I hid the disabled buttons in the demo app, too.

After talking with the designers, we decided to reduce the shadow radius by half to better match the Figma and iOS renderings of the shadows. Some of the shadow definitions weren't quite right, so I updated those. I also checked the rest of our definitions to make sure they match the definitions from the designers. For some reason in the design tokens we use xSmall but xlarge with a lowercase L, so I changed the token definitions to be more consistent and use an uppercase L.