Localizable.stringsdict fix
VoiceOver is reading the hour in the date/time picker control as "%#@hours@". This is due to an issue with the pluralization rule for Accessibility.DateTime.Hour.Value in Localizable.stringsdict: there should be a rule for the variable "hours", but it is under the key "Variable" instead. This change needs to be applied to the Localizable.stringsdict for all languages.