Date/Time Picker: Fix hour and minute accessibility values (#516)
* Localizable.stringsdict fix

VoiceOver is reading the hour in the date/time picker control as "%#@hours@". This is due to an issue with the pluralization rule for Accessibility.DateTime.Hour.Value in Localizable.stringsdict: there should be a rule for the variable "hours", but it is under the key "Variable" instead. This change needs to be applied to the Localizable.stringsdict for all languages.

* Use integer arguments for hour and minute accessibility values

The pluralization rules for Accessibility.DateTime.Hour.Value and Accessibility.DateTime.Minute.Value indicate that their format argument is of type d (integer). However, we're currently supplying strings for these arguments, which get formatted as garbage integers, such as "-1260932919 o'clock". We should pass the integer values directly.