Various changed to the floating button
Changed the check for the additional horizontal padding from state.text?.isEmpty == false to !(state.text?.isEmpty ?? true)

Fixed formatting around isFloatingStyle braces

Removed shadow.none since we're now applying the shadows to the buttons conditionally, so we don't need an "empty" shadow for the regular buttons. Relatedly, changed the default button to use shadow2 since it was the smallest and none is no longer used. Looks like I also forgot to copy the FluentUIStyle.generated.swift over to the actual demo app, so the circleSmall and circleLarge are removed now, also.