Add Floating Action styles to vNext button (#514)
* Add Floating Action styles to vNext button

Added two new styles to the button, accent floating and subtle floating.

We don't have specs for a medium floating button, but the normal button has a medium size, so I made the medium floating buttons mostly use the small button specs and hid them in the demo app. Similarly, the guidance is to not show disabled floating buttons so I hid the disabled buttons in the demo app, too.

After talking with the designers, we decided to reduce the shadow radius by half to better match the Figma and iOS renderings of the shadows. Some of the shadow definitions weren't quite right, so I updated those. I also checked the rest of our definitions to make sure they match the definitions from the designers. For some reason in the design tokens we use xSmall but xlarge with a lowercase L, so I changed the token definitions to be more consistent and use an uppercase L.

* Update vNext button demo controller's background color

Turns out that very dark buttons don't show up too well on very dark backgrounds. Updating the background of the vnext button demo controller to surfaceSecondary allows the subtle floating action buttons to actually show up.

* Address comments not related to text size

Renamed isActive to shouldUsePressedShadow to better reflect the usage of the variable.

Added uses of .modifyIf to setting the additional horizontal padding and setting the shadow to avoid adding zero padding and setting the shadow on styles other than the floating styles.

Since we were checking to see if we were making a floating style, I also changed from using the border radius token to just calling .clipShape(Capsule()), which will handle having semi-circle ends on the button better than adding new radius options.

* Add isFloatingStyle bool to ButtonTokens and make the text size constant

Added a var to MSFButtonStyle to simplify the check for floating style.

Changed the floating styles so that they wouldn't change text size when the user chooses a larger font. Still looking into UILargeContentViewerInteraction alternatives in SwiftUI.

* Various changed to the floating button

Changed the check for the additional horizontal padding from state.text?.isEmpty == false to !(state.text?.isEmpty ?? true)

Fixed formatting around isFloatingStyle braces

Removed shadow.none since we're now applying the shadows to the buttons conditionally, so we don't need an "empty" shadow for the regular buttons. Relatedly, changed the default button to use shadow2 since it was the smallest and none is no longer used. Looks like I also forgot to copy the FluentUIStyle.generated.swift over to the actual demo app, so the circleSmall and circleLarge are removed now, also.

* Move shadow, minimum text height, and additional horizontal padding token definitions to the base button

Since the definitions for the shadow, minimum text height, and additional horizontal padding are shared across both variants of floating buttons, we're moving them to the base button definition to save on code generation and applying them all conditionally.