Avatar border and inner gap improvements (#539)
Currently, when AvatarView shows a border ring, it has an inner gap. Right now, the code is written that white stroke was drawn on top of initialsView or imageView. Instead, adjust the frame size of the content view to show the gap and use UIView layer to draw the border ring.

-updated AvatarGroupView accordingly
-default border color has changed from white to grey
-Border ring was half drawn outside of the avatarSize. Now the border width will be within the avatar bounds.
-.extraSmall size sometimes doesn't render the text appropriately when border is shown. Decrease the minimumScaleFactor
-update borderWidth and innerGap according to design redlines