hide initialsView or Image based on displayStyle of AvatarView (#547)
In case when client team reuses the avatarView, based on the displayStyle hide the subview behind the currentView.
When contactImage was translucent, it may show the initials view behind the current view.

Also, it doesn't make sense just because it doesn't have tooltip text the avatar view isn't accessible. Keep the accessibilityrole to .image if there is a contact image.