hideInsideGapForBorder property in macOS avatarView (#549)
Similar to commit cacd3d8 on iOS, create a new public property hideInsideGapForBorder to hide gap between contentView and borderView

before border view was behind contentView. this can cause weird pixel gaps when we want to draw borderView circle mask adjacent to contentView. When hideInsideGapForBorder is true, make the contentView draw edge to edge of the avatarView and draw the borderView on top