Origin/toddha/updatenavigationbarstyle (#556)
* Make the style property on NavigationBar key-value-observable

* Listen to navigationBarStyle changes in ShyHeaderController

Refactors the old navigationBarStyle code down to a separate function and call from viewDidAppear. Adds an observer so that we update when it changes.

* Add button to test new functionality in the test app

Adds a bunch to change the style periodically which then just does it every 1 second.

We can always set the custom navigation bar color - not based on the style. We just won't use it otherwise.

Bugs I found when testing:
- when entering the search bar, the cancel button doesn't update. We're unlikely to hit this case (change the style while we're in the search view? seems unlikely)
- pill button bar demo doesn't update. This test code would need to listen separately and is outside the scope of this change.

* Pull Request Feedback

Adds comment to explain why we have @objc dynamic - for purposes of KVO
Lint updates
rename changeStyle -> changeStyleContinuously
remove unnecessary self
update newStyle to be immutable once set in the switch statement

Co-authored-by: Todd Harris <toddha@microsoft.com>