[vNext] Fix Avatar image rendering mode
Problem: a while back, I added a forced `.renderingMode(.original) to the `Image` inside `Avatar` to fix an issue with iOS 13. Specifically, when you embed an `Image` inside a `Button` in iOS 13, that `Image` is always forced to a `.template` rendering mode. I manually specified `.original`, and the issue went away for `Avatar`s with an actual image. (#617)

Problem is, sometimes it really *does* need to be rendered with a `renderingMode` of `.template`, depending on which image we're drawing. So this change adds a bit more sophisticated logic to determine which mode to render.