Switch Badge background/text color from NSColor to DynamicColor
The contrast on the dark mode versions of the light mode colors wasn't enough, so we need different colors in light in dark mode. Sounds like a job for our dynamic colors!
Also switched the size enum from public to private, and the colors extension to just a private struct, along with adjusting the corner radius and horizontal padding to match an updated design.