Add simple badge on Mac (#667)
* Add Badge View for Mac

Adding a super simple Badge View for Mac that is only a blue field with blue text. No user interaction or anything.

* Update Badge View Accessibility

Added accessibility label and role, fixed colors not changing between dark and light mode. Also fixed a typo in the demo.

* Add more colors and remove init with size

Since we only have the small size, we don't really need an init where you pass in the size.
Added a new "primary" style, along with making the background and text color accessible to clients so they can have custom colors.

* Switch Badge background/text color from NSColor to DynamicColor

The contrast on the dark mode versions of the light mode colors wasn't enough, so we need different colors in light in dark mode. Sounds like a job for our dynamic colors!
Also switched the size enum from public to private, and the colors extension to just a private struct, along with adjusting the corner radius and horizontal padding to match an updated design.

* Fix podspec file typo

* Make BadgeColors struct private and part of the BadgeView

Renamed the BadgeColors struct Constants, made it private, and moved it into the class.
Updated the corner radius.