Update SearchBar progress spinner light content colors
Search Bar progress spinner should not use the primary tint, but instead use the same as text color in order to ensure that we match accessibility requirements of minimum contrast ratios.

This originally came in as part of commit 4130adbdca2334cf9cc16249ac6e91017008aa8b but does not match the specified designs.