[Improve Usability] BadgeView label text color and other properties need to be an Objc property.
### Environment Information

- **Platform**:
  - [ ] iOS

### Problem:

I am using BadgeView in my project where I can change the text color and background color of the badge in swift file but I cannot do the same for Objc code as label text color and others are not compatible with Objc code.

#### FIX:

Update those properties with @objc

### Priorities and help requested:

Yes - I need this fix to my code as soon as possible as it is blocking my feature.

Products/applications affected: (if applicable)
Microsoft Office.

#### VALIDATION:

-- verify the build and it is green.