Bring vNext button to Main
Brought over the vNext button, along with related changes, from the vNext branch. Most of these changes came straight from there, but the exception is the Objective C demo controller. I didn't bring over the Objective C demo from vNext because it still uses the old button, and instead rewrote it.

Also, I updated the TabBarView demo controller +/- buttons to use icons instead of text to match the rest of the +/- buttons, and updated ic_fluent_subtract_20_regular to be a template image.