Add subspec for Divider
Add a subspec for the divider to MicrosoftFluentUIVnext.podspec

Remove background color MSFDivider's view and add a constraint to set the height or width to one device pixel if the divider is horizontal or vertical respectively.

Added margins to the vertical stack in the demo to make it easier to actually see the dividers while not adding spacing to the stack that would be confused for the spacing of the dividers.
Added text between the vertical dividers in the demo
Fixed some ambiguity in the horizontal stack by setting the distribution to equal centering