Add Fluent Divider (#803)
* Add tokenized Divider

Add a new SwiftUI Separator that uses the token system. Also, change the name from separator to divider.

* Address various comments from the first iteration

Moved the new Divider under the vNext folder
Separated out MSFDivider
Removed switch from body function in favor of ternary operators
Added custom color (black) demo
Switched from width and height of frame to maxWidth and maxHeight
Removed shadow style (which then means removing MSFDividerStyle in general since now there's only one)
Added missing spaces around {} and added new lines to longer parameter lists
Added initializer to MSFDivider with theme
Added .designTokens modifier to FluentDivider

* Add Swift UI demo

Added a Swift UI demo, along with changing the existing demo controller to use UITableViewController rather than DemoController

Changed the custom color demo to communication blue so it would have dark mode support

Fixed the tokens looking for the .rest instead of just the color

Fixed MSFDivider trying to recursively return its own window rather than its view's window

* Add subspec for Divider

Add a subspec for the divider to MicrosoftFluentUIVnext.podspec

Remove background color MSFDivider's view and add a constraint to set the height or width to one device pixel if the divider is horizontal or vertical respectively.

Added margins to the vertical stack in the demo to make it easier to actually see the dividers while not adding spacing to the stack that would be confused for the spacing of the dividers.
Added text between the vertical dividers in the demo
Fixed some ambiguity in the horizontal stack by setting the distribution to equal centering

* Align SwiftUI Demo with other demos

Changed the SwiftUI demo to have 5 dividers and two strings to match the UIKit demo and better show the divider in action, rather than just a line

Add padding to height constraint of the hosting UIView to bring back the missing medium dividers

* Add a call in MSFDivider to disable Safe Area insets on the hosting controller

Add a call in MSFDivider to disable Safe Area insets on the hosting controller