Update usage of MSFSeparator and SwiftUI Divider to Fluent Divider
Updated most instances of either the old separator or the SwiftUI divider to the new Fluent Divider.

Currently running in to an issue with Fluent Dividers that are used in UITableViewCell where they'll either have a constraint issue that is broken by the system, or if I try to set translatesAutoresizingMaskIntoConstraints to false for the divider the hosting cell gets stuck calling layoutSubviews.

Also, got approval from the designers to update the color from netural1 to neutral2 to increase contrast