Update usage of MSFSeparator and SwiftUI Divider to Fluent Divider (#833)
* Add tokenized Divider

Add a new SwiftUI Separator that uses the token system. Also, change the name from separator to divider.

* Address various comments from the first iteration

Moved the new Divider under the vNext folder
Separated out MSFDivider
Removed switch from body function in favor of ternary operators
Added custom color (black) demo
Switched from width and height of frame to maxWidth and maxHeight
Removed shadow style (which then means removing MSFDividerStyle in general since now there's only one)
Added missing spaces around {} and added new lines to longer parameter lists
Added initializer to MSFDivider with theme
Added .designTokens modifier to FluentDivider

* Add Swift UI demo

Added a Swift UI demo, along with changing the existing demo controller to use UITableViewController rather than DemoController

Changed the custom color demo to communication blue so it would have dark mode support

Fixed the tokens looking for the .rest instead of just the color

Fixed MSFDivider trying to recursively return its own window rather than its view's window

* Add subspec for Divider

Add a subspec for the divider to MicrosoftFluentUIVnext.podspec

Remove background color MSFDivider's view and add a constraint to set the height or width to one device pixel if the divider is horizontal or vertical respectively.

Added margins to the vertical stack in the demo to make it easier to actually see the dividers while not adding spacing to the stack that would be confused for the spacing of the dividers.
Added text between the vertical dividers in the demo
Fixed some ambiguity in the horizontal stack by setting the distribution to equal centering

* Align SwiftUI Demo with other demos

Changed the SwiftUI demo to have 5 dividers and two strings to match the UIKit demo and better show the divider in action, rather than just a line

Add padding to height constraint of the hosting UIView to bring back the missing medium dividers

* Add a call in MSFDivider to disable Safe Area insets on the hosting controller

Add a call in MSFDivider to disable Safe Area insets on the hosting controller

* Update usage of MSFSeparator and SwiftUI Divider to Fluent Divider

Updated most instances of either the old separator or the SwiftUI divider to the new Fluent Divider.

Currently running in to an issue with Fluent Dividers that are used in UITableViewCell where they'll either have a constraint issue that is broken by the system, or if I try to set translatesAutoresizingMaskIntoConstraints to false for the divider the hosting cell gets stuck calling layoutSubviews.

Also, got approval from the designers to update the color from netural1 to neutral2 to increase contrast

* Removing constraints from UIKit wrapper, adding frame parameters in the SwiftUI layer.

* Move Thickness to the divider state

In order to access the thickness from outside the divider, I've moved it to the state. This lets us correctly set up the TableViewCell

Also swapped the SwiftUI Divider in the vNext List to the FluentDivider

Cleaned up some unneeded translatesAutoresizingMaskIntoConstraints and comments

* Replace closing curly brace that got eaten by my merge

It would seem I deleted a bit too much with my merge. Adding back a }

* Hold on to dividers and other comments

Added non-local variable to hold on to the dividers for the life time of the hosting view

Added constraints to the UIKit divider demo to fix ambiguous height/width warnings

Removed .overlays on FluentDivider from List and ListCell. Relatedly, removed ListCell's border color token

Added local variables to reduce repeated instances of property access

* Update Pod Specs

Update the pod specs to depend on the Divider rather than the Separator

* Change repetitive array inits to map

Changed arrays inited with multiple dividers to use map rather than just initing with multiple MSFDividers.

Fixed spacing in Divider.swift

Fixed Divider podspec

* Move TableViewCell dividers to contentView

Move the dividers to the TableViewCell's content view rather than adding them as direct subviews. Also move the layout calls to layoutContentSubviews.

* Replace Divider Demo constraint with stack distribution change

Changing the stack to equal centering allows the stack to deal with the ambiguous heights by itself.

* Change new range operators to half-open

There will be new guidance in our swift guide that says we prefer half-open range operators, so changing the new ones I added to match the upcoming guidance

Co-authored-by: Rafael Amaral de Assis <rdeassis@microsoft.com>