[vNext Tokens] Convert `FluentDivider`  (#874)
* Add new DividerTokens

Add new Divider Tokens based on the new token system

* Update Divider to use the new tokens

Updated usage of old MSFDividerTokens to new DividerTokens

* Remove generated Divider tokens from the Fluent project

Removed the generated tokens now that the new tokens are working

* Removed Divider tokens from SGen

Removed the Divider tokens from SGen so they wouldn't be generated anymore

* Move overrideTokens to DividerModifiers.swift

Added DividerModifiers.swift and moved the implementation overrideTokens there.

* Update MSFDividerStateImpl to init tokens with spacing

Updated MSFDividerStateImpl to init the defaultTokens using the passed in spacing, making so we don't need a default value in the tokens anymore.

* Remove color from DividerState

With the new token system, we should be pulling the color from the tokens rather than the state. Removed color from the state and added override tokens to replace usages of state.color. Also had to make our dynamicColor extension of UIColor public

* Change resolve modifier from orientation to spacing

I think I misunderstood .resolveTokenModifier when I first added it, but I don't think it should be based on orientation. Spacing makes more sense.

* Update Divider demo to reduce how often tokens are overridden

Changed the way we create dividers in the divider demo to only override tokens when we're trying to use a custom color.

* Update MSFDivider to ControlHostingContainer

In order for the new token system to work, MSFDivider also needed an update to be a ControlHostingContainer. Now the custom tokens are actually applied in the demo.

* Remove MSFDivider init with FluentUIStyle param

We no longer need the FluentUIStyle param with the new token system, so simplifying MSFDivider to have just the one init.

* Add comment to overrideTokens in FluentDivider extension

* Add comment to dynamicColor in UIColor extension

* Move dynamicColor higher in UIColor+Extensions.swift

Moved dynamicColor above the internal inits to match our Swift coding guidelines. (Though I guess git is reading it as moving the internal inits down)

* Update MSFDivider to match Swift Guidelines

Moved the state below the init to match our guidelines.