Add largeContentViewer support
To support large accessibility text sizes, larger content viewer support as been added. The card view has dynamic text support which causes the height of the card to change. The width of the card will not be changing though so that when using in a layout like a collection view we dont get different grid layout based on dynamic text.