[vNext Tokens] Convert PillButton to the new token system (#903)
* Add new PillButton tokens

Add new tokens for the PillButton using the new token system. Also added PillButtonDynamicColors in the style of ButtonDynamicColors. Which also included adding a hover title color.

* Update PillButton to conform to ControlConfiguration

Updated PillButton to conform to ControlConfiguration protocol, along with stub of updatePillButtonTokens that'll get its body uncommented once I update to TokenizedControlInternal.

* Workaround for collision with UIControl.state

PillButton is a UIButton, so it's a UIControl, and those have states. So we can't just override the state of TokenizedControlInternal like we normally would. Add in an alternative to get me unblocked while I figure everything else out.

* Update PillButton to TokenizedUIControlInternal (emphasis on the UI)

* Update PillButton to not be FluentUIWindowProvider

* Update PillButton to use the new tokens

* Remove old tokens from Fluent project

* Remove old PillButton tokens from sgen

* Reorder file to match Swift guidelines

* Remove an extra line

* Set style in token update

For some reason the token resolver was just handing back the default tokens, and we weren't getting the tokens for the correct style.

* Move custom colors to tokens

Also had to add a public init to the PillButtonDynamicColors, and make the lets public.

* Remove custom colors from Pill Button Bar (and extra inits)

* Add call to updateAppearance when setting overrideTokens

* Remove token type typealias from PillButton

* Reduce colors of PillButtonDynamicColors

Hover and highlight variants aren't really spec'd out, or different from other colors, so they're being removed.