[vNext Tokens] Convert PillButtonBar to the new token system (#911)
* Add new tokens for PillButtonBar

* Update PillButtonBar to conform to ControlConfiguration

With a stub that'll be updated once PillButtonBar also conforms to TokenizedControlInternal

* Update PillButtonBar to conform to TokenizedControlInternal

I had to add the TokenType typealias again. I was able to remove it with the PillButton, but it doesn't want to build without it. Will revisit and see if there was some other change later that helped get rid of it.

* Reorder PillButtonBar.swift to better match our Swift guidelines

* Fix trailing whitespace

* Update PillButtonBar to no longer conform to FluentUIWindowProvider

* Update PillButtonBar to use the new tokens

Also had to rename a few different tokens

* Remove old PillButtonBar tokens from Fluent project

* Remove PillButtonBar tokens from SGen

* Update PillButtonBar to conform to updated TokenizedControl

Don't need state anymore, now need updateCurrentTokens

* Update PillButton/Bar and CommandBar to no longer conform to ControlConfiguration