Update SegmentedControl to support images
Added an optional image to the SegmentedItem
Changed pillMaskedLabelsContainerView's name to pillMaskedContentContainerView, since now it holds more than just labels.
Changed the maskedLabels array to be an array of optionals
Also updated the logic in sizeThatFits, since I was getting weird results where the window's layout guide was nil or something and just returning zero.
Moved some logic out of for loops