Replace usages of fatalError(...) with preconditionFailure(...) or assertionFailure(...) where applicable
Problem:
When fatalError is called in both Debug and Release builds, it will print the full system path of the source file. This behvaior can be considered undesirable for Release builds running in production environments.

https://lapcatsoftware.com/articles/fatalError.html

Fix:
Replace usages of fatalError(...) with preconditionFailure(...). preconditionFailure(...) was used in all cases here so that apps will continue to terminate if the function is called in a release build. assertionFailure(...)
will not terminate the app if called in a release build.

Verification:
- Built FluentUI static library.
- Verified functionality in respective Demo controllers
- Verified objects initialize as expected changes do not cause unwanted behavior.